adapt code
