+++ web/src/js/components/FlowView/Messages.tsx
@@ -1,7 +1,6 @@
 import { Flow, MessagesMeta } from "../../flow";
 import { useAppDispatch, useAppSelector } from "../../ducks";
 import * as React from "react";
-import { useCallback, useState } from "react";
 import {
     ContentViewData,
     useContentView,
@@ -9,7 +8,7 @@ import {
 import ViewSelector from "../contentviews/ViewSelector";
 import { setContentViewFor } from "../../ducks/ui/flow";
 import { formatTimeStamp } from "../../utils";
-import ContentRenderer from "../contentviews/ContentRenderer";
+import ContentEditor from "../contentviews/ContentEditor";
 
 type MessagesPropTypes = {
     flow: Flow;
@@ -22,23 +21,13 @@ export default function Messages({ flow, messages_meta }: MessagesPropTypes) {
     const contentView = useAppSelector(
         (state) => state.ui.flow.contentViewFor[flow.id + "messages"] || "Auto",
     );
-    const [maxLines, setMaxLines] = useState<number>(
-        useAppSelector((state) => state.options.content_view_lines_cutoff),
-    );
-    const showMore = useCallback(
-        () => setMaxLines(Math.max(1024, maxLines * 2)),
-        [maxLines],
-    );
     const messages =
         useContentView(
             flow,
             "messages",
             contentView,
-            maxLines + 1,
             flow.id + messages_meta.count,
         ) ?? [];
-
-    let remainingLines = maxLines;
 
     return (
         <div className="contentview">
@@ -68,14 +57,13 @@ export default function Messages({ flow, messages_meta }: MessagesPropTypes) {
                                 {d.timestamp && formatTimeStamp(d.timestamp)}
                             </span>
                         </small>
-                        <ContentRenderer
-                            content={d.text}
-                            maxLines={remainingLines}
-                            showMore={showMore}
+                        <ContentEditor
+                            initialContent={d.text}
+                            language={d.syntax_highlight}
+                            readonly
                         />
                     </div>
                 );
-                remainingLines -= d.text.split("\n").length;
                 return renderer;
             })}
         </div>
