+++ web/src/js/components/contentviews/ContentEditor.tsx
@@ -5,47 +5,61 @@ import { javascript } from "@codemirror/lang-javascript";
 import { yaml } from "@codemirror/lang-yaml";
 import { css } from "@codemirror/lang-css";
 import { html } from "@codemirror/lang-html";
+import { SyntaxHighlight } from "../../backends/consts";
 
-type ContentEditorProps = {
-    content: string;
-    onChange: (content: string) => void;
+type CodeEditorProps = {
+    initialContent: string;
+    onChange?: (content: string) => void;
     readonly?: boolean;
-    language?: string;
+    language?: SyntaxHighlight | null;
 };
 
-export default function ContentEditor({
-    content,
+export default function CodeEditor({
+    initialContent,
     onChange,
     language,
     readonly = false,
-}: ContentEditorProps) {
+}: CodeEditorProps) {
     const stopPropagation = useCallback(
         (e: React.KeyboardEvent<HTMLDivElement>) => e.stopPropagation(),
         [],
     );
     const extensions = useMemo(() => {
         switch (language) {
-            case "javascript":
-                return [javascript()];
-            case "yaml":
+            case SyntaxHighlight.YAML:
                 return [yaml()];
-            case "css":
+            case SyntaxHighlight.XML:
+                return [html()];
+            case SyntaxHighlight.JAVASCRIPT:
+                return [javascript()];
+            case SyntaxHighlight.CSS:
                 return [css()];
-            case "html":
-                return [html()];
-            default:
-                return [html()];
+            case undefined:
+            case null:
+            case SyntaxHighlight.NONE:
+            case SyntaxHighlight.ERROR:
+                return [];
+            /* istanbul ignore next @preserve */
+            default: {
+                const unexpected: never = language;
+                console.error(
+                    "Unexpected syntax highlighting language: ",
+                    unexpected,
+                );
+                return [];
+            }
         }
     }, [language]);
-    
     return (
         <div className="codeeditor" onKeyDown={stopPropagation}>
             <CodeMirror
-                value={content}
+                value={initialContent}
                 onChange={onChange}
                 readOnly={readonly}
                 extensions={extensions}
-                basicSetup={{ highlightActiveLine: false }}
+                basicSetup={{
+                    highlightActiveLine: false,
+                }}
             />
         </div>
     );
