+++ web/src/js/components/contentviews/useContentView.tsx
@@ -2,11 +2,12 @@ import { useMemo } from "react";
 import { useContent } from "./useContent";
 import { Flow, HTTPFlow, HTTPMessage } from "../../flow";
 import { MessageUtils } from "../../flow/utils";
+import { SyntaxHighlight } from "../../backends/consts";
 
 export type ContentViewData = {
     text: string;
     view_name: string;
-    syntax_highlight: string;
+    syntax_highlight: SyntaxHighlight;
     description: string;
     from_client?: boolean;
     timestamp?: number;
@@ -16,7 +17,6 @@ export function useContentView(
     flow: Flow,
     part: "messages",
     view?: string,
-    lines?: number,
     hash?: string,
 ): ContentViewData[] | undefined;
 
@@ -24,7 +24,6 @@ export function useContentView(
     flow: HTTPFlow,
     part: HTTPMessage | "request" | "response",
     view?: string,
-    lines?: number,
     hash?: string,
 ): ContentViewData | undefined;
 
@@ -32,10 +31,9 @@ export function useContentView(
     flow: Flow,
     part: HTTPMessage | "request" | "response" | "messages",
     view?: string,
-    lines?: number,
     hash?: string,
 ): ContentViewData | ContentViewData[] | undefined {
-    const url = MessageUtils.getContentURL(flow, part, view, lines);
+    const url = MessageUtils.getContentURL(flow, part, view);
     const cv_json = useContent(url, hash);
     return useMemo<ContentViewData | undefined>(() => {
         if (cv_json) {
@@ -45,7 +43,7 @@ export function useContentView(
                 const err: ContentViewData = {
                     text: cv_json,
                     description: "Network Error",
-                    syntax_highlight: "error",
+                    syntax_highlight: SyntaxHighlight.ERROR,
                     view_name: "raw",
                 };
                 if (part === "messages") {
